+++ src/collector/index.ts
@@ -1,23 +1,23 @@
-import { IntegrationInstance } from '@jupiterone/integration-sdk-core';
+import {
+  IntegrationExecutionContext,
+  IntegrationValidationError,
+} from '@jupiterone/integration-sdk-core';
 import { ServicesClient } from './ServicesClient';
 import { CloudflareIntegrationConfig } from '../types';
-
-export * from './types';
 
 /**
  * Creates a ServicesClient from an integration instance using it's
  * api key.
  */
 export function createServicesClient(
-  instance: IntegrationInstance<CloudflareIntegrationConfig>,
+  context: IntegrationExecutionContext<CloudflareIntegrationConfig>,
 ): ServicesClient {
-  const { apiToken } = instance.config;
+  const { apiToken } = context.instance.config;
 
   if (!apiToken) {
-    throw new Error(
+    throw new IntegrationValidationError(
       'Required configuration item "apiToken" is missing on the integration instance config',
     );
   }
-
-  return new ServicesClient({ apiToken });
+  return new ServicesClient(context.instance.config, context.logger);
 }