+++ CHANGELOG.md
@@ -8,6 +8,16 @@ and this project adheres to
 
 ## [Unreleased]
 
+## 0.4.0 - 2020-10-22
+
+### Changed
+
+- Changed `cloudflare_account_role|assigned|member` to
+  `cloudflare_account_member|assigned|role`
+- Upgrade sdk@v3.6.0
+- Use iterators in client; don't hold all entity data in memory
+- Use `@cloudflare/types` typing instead of internal types
+
 ## 0.1.0 - 2020-05-18
 
 ### Added